+++ duckscript_sdk/src/sdk/std/eval/mod.rs
@@ -37,7 +37,7 @@ impl Command for CommandImpl {
         _output_variable: Option<String>,
         _instructions: &Vec<Instruction>,
         commands: &mut Commands,
-        _test_setline: usize,
+        _line: usize,
     ) -> CommandResult {
         eval::eval(&arguments, state, variables, commands)
     }